Modification API pour intégration sécurisation JWT Token

+++ api_auth/main.py
@@ -1,12 +1,12 @@
 from datetime import datetime, timedelta
 from typing import Optional
 from fastapi import FastAPI, HTTPException, Depends, status
-from fastapi.security import OAuth2PasswordBearer
 import mysql.connector
 from pydantic import BaseModel
 import psycopg2
 from psycopg2.extras import RealDictCursor
 from jose import JWTError, jwt
+from fastapi import Header
 
 SECRET_KEY = "sae501"
 ALGORITHM = "HS256"
@@ -31,12 +31,27 @@ def create_access_token(data: dict):
     encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
     return encoded_jwt
 
-def getIdFournisseur(fournisseur_nom: str, db_conn):
-    with db_conn.cursor(cursor_factory=RealDictCursor) as cursor:
-        cursor.execute("SELECT NoFournisseur FROM Fournisseur WHERE NomFournisseur = %s", (fournisseur_nom,))
-        result = cursor.fetchone()
-        return result['NoFournisseur'] if result else None
-
+async def get_current_user(token: str = Header(None, alias="token"), db_conn: mysql.connector.connection.MySQLConnection = Depends(get_db_conn_mysql)):
+    credentials_exception = HTTPException(
+        status_code=status.HTTP_401_UNAUTHORIZED,
+        detail="Impossible de valider les informations d'identification",
+        headers={"WWW-Authenticate": "Bearer"},
+    )
+    if token is None:
+        raise credentials_exception
+    try:
+        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
+        username: str = payload.get("sub")
+        if username is None:
+            raise credentials_exception
+        cursor = db_conn.cursor(dictionary=True)
+        cursor.execute("SELECT login, statut FROM comptes WHERE login = %s", (username,))
+        user = cursor.fetchone()
+        if user is None:
+            raise credentials_exception
+        return user
+    except JWTError:
+        raise credentials_exception
 
 # Modèle Pydantic pour la création et la mise à jour d'un compte
 class CompteCreate(BaseModel):
@@ -50,10 +65,10 @@ class CompteUpdate(BaseModel):
         
 # Opération CRUD : Lire un compte par login
 @app.get("/comptes/{login}", response_model=dict)
-async def read_compte(login: str, db_conn: mysql.connector.connection.MySQLConnection = Depends(get_db_conn_mysql)):
+async def read_compte(login: str, current_user: dict = Depends(get_current_user), db_conn: mysql.connector.connection.MySQLConnection = Depends(get_db_conn_mysql)):
     cursor = db_conn.cursor(dictionary=True)
     try:
-        query = "SELECT login, password, statut FROM comptes WHERE login = %s"
+        query = "SELECT login, statut FROM comptes WHERE login = %s"
         cursor.execute(query, (login,))
         compte = cursor.fetchone()
         if compte:
@@ -62,6 +77,7 @@ async def read_compte(login: str, db_conn: mysql.connector.connection.MySQLConne
             raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Compte introuvable")
     finally:
         cursor.close()
+
         
 # Route pour vérifier l'existence d'un compte avec des données JSON
 @app.post("/check_account")
@@ -88,7 +104,12 @@ async def check_account(data: dict, db_conn: mysql.connector.connection.MySQLCon
 
 # Opération CRUD : Créer un compte
 @app.post("/comptes/", response_model=dict)
-async def create_compte(compte: CompteCreate, db_conn: mysql.connector.connection.MySQLConnection = Depends(get_db_conn_mysql)):
+async def create_compte(compte: CompteCreate, current_user: dict = Depends(get_current_user), db_conn: mysql.connector.connection.MySQLConnection = Depends(get_db_conn_mysql)):
+    # Vérifier si l'utilisateur actuel a les autorisations nécessaires pour créer un compte
+    # Par exemple, vérifier si le statut de l'utilisateur est 'admin'
+    if current_user["statut"] != "administrateur":
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Opération non autorisée")
+
     cursor = db_conn.cursor(dictionary=True)
     try:
         query = "INSERT INTO comptes (login, password, statut) VALUES (%s, %s, %s)"
@@ -96,13 +117,20 @@ async def create_compte(compte: CompteCreate, db_conn: mysql.connector.connectio
         db_conn.commit()
         return {"status": "success", "message": "Compte créé avec succès", "compte": compte.dict()}
     except mysql.connector.Error as e:
+        db_conn.rollback()  # Assurez-vous d'annuler la transaction en cas d'erreur
         return {"status": "error", "message": f"Erreur lors de la création du compte : {e}"}
     finally:
         cursor.close()
 
+
 # Opération CRUD : Mettre à jour un compte par login
 @app.put("/comptes/{login}", response_model=dict)
-async def update_compte(login: str, compte_update: CompteUpdate, db_conn: mysql.connector.connection.MySQLConnection = Depends(get_db_conn_mysql)):
+async def update_compte(login: str, compte_update: CompteUpdate, current_user: dict = Depends(get_current_user), db_conn: mysql.connector.connection.MySQLConnection = Depends(get_db_conn_mysql)):
+    # Vérifier si l'utilisateur actuel a les autorisations nécessaires pour mettre à jour le compte
+    # L'utilisateur peut mettre à jour son propre compte ou doit être un admin pour mettre à jour les comptes d'autres utilisateurs
+    if current_user["statut"] != "administrateur":
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Opération non autorisée")
+
     cursor = db_conn.cursor(dictionary=True)
     try:
         query = "UPDATE comptes SET password = %s, statut = %s WHERE login = %s"
@@ -110,13 +138,20 @@ async def update_compte(login: str, compte_update: CompteUpdate, db_conn: mysql.
         db_conn.commit()
         return {"status": "success", "message": "Compte mis à jour avec succès"}
     except mysql.connector.Error as e:
+        db_conn.rollback()  # Assurez-vous d'annuler la transaction en cas d'erreur
         return {"status": "error", "message": f"Erreur lors de la mise à jour du compte : {e}"}
     finally:
         cursor.close()
 
+
 # Opération CRUD : Supprimer un compte par login
 @app.delete("/comptes/{login}", response_model=dict)
-async def delete_compte(login: str, db_conn: mysql.connector.connection.MySQLConnection = Depends(get_db_conn_mysql)):
+async def delete_compte(login: str, current_user: dict = Depends(get_current_user), db_conn: mysql.connector.connection.MySQLConnection = Depends(get_db_conn_mysql)):
+    # Vérifier si l'utilisateur actuel a les autorisations nécessaires pour supprimer le compte
+    # L'utilisateur peut supprimer son propre compte ou doit être un admin pour supprimer les comptes d'autres utilisateurs
+    if current_user["statut"] != "administrateur":
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Opération non autorisée")
+
     cursor = db_conn.cursor(dictionary=True)
     try:
         query = "DELETE FROM comptes WHERE login = %s"
@@ -124,9 +159,11 @@ async def delete_compte(login: str, db_conn: mysql.connector.connection.MySQLCon
         db_conn.commit()
         return {"status": "success", "message": "Compte supprimé avec succès"}
     except mysql.connector.Error as e:
+        db_conn.rollback()  # Assurez-vous d'annuler la transaction en cas d'erreur
         return {"status": "error", "message": f"Erreur lors de la suppression du compte : {e}"}
     finally:
         cursor.close()
+
 
 if __name__ == "__main__":
     import uvicorn
